Collapse dynamic LRM into enhanced LRM
As dynamic is trivial, and NOOP as default results
in enhanced to behave as today (as before), so this
is fine for existing users as well.
